Fix bf16 crash in predict_action numpy conversion
numpy cannot represent BFloat16, so serving with --use_bf16 made every
inference request fail at the normalized_actions conversion. Cast to
float32 first, matching the embodied_action_tokens handling.

+++ starVLA/model/framework/VLA_JEPA.py
@@ -589,7 +589,9 @@ class VLA_JEPA(baseframework):
             key: value.detach() for key, value in diagnostics.items()
         } if diagnostics else None
         public_output = {
-            "normalized_actions": pred_actions.detach().cpu().numpy(),
+            "normalized_actions": pred_actions.to(
+                dtype=torch.float32
+            ).detach().cpu().numpy(),
             "embodied_action_tokens": qwen.embodied_action_tokens.to(
                 dtype=torch.float32
             ).detach().cpu().numpy(),
